move serializers to sentry_apps
