+++ src/sentry/api/serializers/models/__init__.py
@@ -4,7 +4,6 @@ from .apiapplication import *  # noqa: F401,F403
 from .apiauthorization import *  # noqa: F401,F403
 from .apikey import *  # noqa: F401,F403
 from .apitoken import *  # noqa: F401,F403
-from .app_platform_event import *  # noqa: F401,F403
 from .auditlogentry import *  # noqa: F401,F403
 from .auth_provider import *  # noqa: F401,F403
 from .broadcast import *  # noqa: F401,F403
@@ -37,7 +36,6 @@ from .organization_member.scim import *  # noqa: F401,F403
 from .organization_member.utils import *  # noqa: F401,F403
 from .organization_plugin import *  # noqa: F401,F403
 from .orgauthtoken import *  # noqa: F401,F403
-from .platformexternalissue import *  # noqa: F401,F403
 from .plugin import *  # noqa: F401,F403
 from .project import *  # noqa: F401,F403
 from .project_key import *  # noqa: F401,F403
@@ -57,11 +55,6 @@ from .repository import *  # noqa: F401,F403
 from .role import *  # noqa: F401,F403
 from .rule import *  # noqa: F401,F403
 from .savedsearch import *  # noqa: F401,F403
-from .sentry_app import *  # noqa: F401,F403
-from .sentry_app_avatar import *  # noqa: F401,F403
-from .sentry_app_component import *  # noqa: F401,F403
-from .sentry_app_installation import *  # noqa: F401,F403
-from .servicehook import *  # noqa: F401,F403
 from .tagvalue import *  # noqa: F401,F403
 from .team import *  # noqa: F401,F403
 from .user_social_auth import *  # noqa: F401,F403
