+++ src/sentry/sentry_apps/api/serializers/app_platform_event.py
