+++ src/sentry/sentry_apps/api/serializers/platform_external_issue.py
