+++ src/sentry/sentry_apps/api/serializers/sentry_app_component.py
