+++ src/sentry/sentry_apps/api/serializers/sentry_app_installation_response.py
