+++ tests/sentry/api/serializers/test_sentry_app.py
@@ -1,9 +1,9 @@
 from datetime import datetime, timedelta
 
 from sentry.api.serializers import serialize
-from sentry.api.serializers.models.sentry_app import SentryAppSerializer
 from sentry.auth import access
 from sentry.models.avatars.sentry_app_avatar import SentryAppAvatar
+from sentry.sentry_apps.api.serializers.sentry_app_response import SentryAppSerializer
 from sentry.testutils.cases import TestCase
 from sentry.testutils.helpers.datetime import freeze_time
 from sentry.testutils.silo import control_silo_test, no_silo_test
